Centroid tracking
This is a simple tracking strategy that considers that the players just move a
bit from a frame to the next. The simplified steps are as follow:

1. Detect the players on the first frame
2. Save their centroids and give them an ID
3. For the following frames:
   a. Detect the players
   b. For each player compare his centroid with the old centroids
      i. If the distance is small, consider him to be the same player
      ii. Else, give him a new ID

As I said, this is a simples solution. It gives us a nice result for basic tasks
but is very sensible. For example, if a player was not detected in a frame, but
was detected in the frames before and after, a new ID will be created anyway.

I'm aware of this kind of problem, and I expect to solve them post-processing
the detection result.